Only write indicate or notify to CCCD on Android
Some peripherals have characteristics that support either notification
or indication, when we're observation such a characteristic, we should
write to the Client Characteristic Configuration descriptor (CCCD)
either to enable notifications **or** indications, **not** both.

We default to **notifications** if characteristic supports both. This is
consistent with the documented behavior of [Core Bluetooth].

[Core Bluetooth]: https://developer.apple.com/documentation/corebluetooth/cbperipheral/1518949-setnotifyvalue