Make `services` direct references and match properties for I/O
Previously, all I/O operations would lazily search for the first match
when looking up characteristics or descriptors. This was problematic if
duplicate UUIDs existed. Discovered GATT profile items (services,
characteristics and descriptors) as provided via `Peripheral.services`
are now direct references to underlying platform types. This means they
no longer trigger lookups for I/O operations and specific GATT profile
items can be sought after and used.

When using `characteristicOf`, the properties of the characteristic are
now taken into account when searching for the underlying platform
object. For example, when performing a read, only a characteristic with
a matching UUID **and** `read` property will be considered a match.