Revert uuid to v0.6.0
We're hitting the following build failure:

```
> Task :bridge:compileKotlinJs FAILED
e: Could not find "uuid" in [/home/runner/.local/share/kotlin/daemon]
e: java.lang.IllegalStateException: FATAL ERROR: Could not find "uuid" in [/home/runner/.local/share/kotlin/daemon]
```

We don't suspect this to be directly caused by `uuid` library, but
rather a Kotlin version incompatibility that for some reason surfaces
when this library is upgraded. We're continuing to investigate the cause
but this will unblock us for now.